Move emoji modules
