+++ packages/livechat/src/components/uiKit/message/Mrkdwn/index.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'preact/compat';
 
 import { createClassName } from '../../../../helpers/createClassName';
-import shortnameToUnicode from '../../../Emoji/shortnameToUnicode';
+import shortnameToUnicode from '../../../../lib/emoji/shortnameToUnicode';
 import MarkdownBlock from '../../../MarkdownBlock';
 import styles from './styles.scss';
