+++ packages/livechat/src/lib/emoji/ascii.ts
