+++ packages/livechat/src/lib/emoji/isBigEmoji.ts
